feat: integrate MegaMoe fused MoE for DSV4 Python path
Adds the aclnnMegaMoe fused dispatch+GEMM+combine kernel as an alternative
to the grouped-MoE + EP all-reduce path (--enable_mega_moe=true, W8A8
quantization via int64-encoded weight scales).

C++ side:
- npu_ops_library: register the mega_moe TORCH_LIBRARY schema + PrivateUse1
  impl (float schema / double C++ signature per ATen convention)
- npu_process_group: save host/port/group metadata at construction;
  acquire_mega_moe_hccl_comm() creates a DEDICATED HCCL communicator
  (HcclGetRootInfo -> TCP-store rendezvous -> HcclCommInitRootInfo) --
  the c10d-managed comm's op-resource-context conflicts with
  HcclCreateOpResCtx (HCCL_E_INTERNAL result=5)
- py_causal_lm: mega_moe_context_tensor/ccl_buffer_size bridge using the
  dedicated comm; config_dict passes enable_mega_moe to Python
- py_executor_impl: expose mega_moe_context_tensor / mega_moe_ccl_buffer_size
  via xllm_runtime pybind

Python side:
- kernels_npu/moe.py: mega_moe wrapper + encode_w8a8_scale_int64
  (fp32 -> view(int32) -> int64 bit-cast) + has_mega_moe
- deepseek_v4.py: DeepseekV4MoE._forward_mega_moe (global topk routing,
  per-expert TensorList weights, int64 scales, dispatch_quant_mode=2)
  gated by enable_mega_moe && ep_size > 1

Tests: test_mega_moe.py covers scale encoding bit-cast, config plumbing,
and model branch selection (8/10 pass, 2 need NPU runtime).

Status: HCCL comm initialization and HcclCreateOpResCtx both succeed.
End-to-end validation blocked by external processes occupying cards
(pending verification when cards free).

+++ xllm/core/kernels/npu/npu_mega_moe.cpp
@@ -57,7 +57,7 @@ std::tuple<torch::Tensor, torch::Tensor> apply_npu_mega_moe(
     const std::string& comm_alg,
     int64_t num_max_tokens_per_rank,
     const std::string& activation,
-    float activation_clamp,
+    double activation_clamp,
     int64_t dispatch_quant_out_dtype,
     int64_t topo_type,
     int64_t rank_num_per_server) {
